+++ src/Blitz.php
@@ -24,8 +24,6 @@ use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterUserPermissionsEvent;
 use craft\helpers\UrlHelper;
 use craft\log\MonologTarget;
-use craft\queue\jobs\ResaveElements;
-use craft\queue\Queue as CraftQueue;
 use craft\services\Dashboard;
 use craft\services\Elements;
 use craft\services\Plugins;
@@ -67,7 +65,6 @@ use yii\base\Event;
 use yii\di\Instance;
 use yii\log\Dispatcher;
 use yii\log\Logger;
-use yii\queue\ExecEvent;
 use yii\queue\Queue;
 
 /**
@@ -119,7 +116,7 @@ class Blitz extends Plugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '5.6.0';
+    public string $schemaVersion = '5.7.0.1';
 
     /**
      * @inheritdoc
@@ -146,17 +143,24 @@ class Blitz extends Plugin
         $this->registerVariables();
         $this->registerLogTarget();
 
+        // Potentially refresh the cache at the end of every request
+        Craft::$app->onAfterRequest(function() {
+            $this->refreshCache->refresh();
+        });
+
         // Register events
         $this->registerCacheableRequestEvents();
-        $this->registerElementEvents();
-        $this->registerResaveElementEvents();
-        $this->registerStructureEvents();
-        $this->registerIntegrationEvents();
-        $this->registerHintsUtilityEvents();
         $this->registerClearCaches();
+        if ($this->settings->refreshCacheEnabled) {
+            $this->registerElementEvents();
+            $this->registerStructureEvents();
+            $this->registerIntegrationEvents();
+        }
 
-        // Register control panel events
-        if (Craft::$app->getRequest()->getIsCpRequest()) {
+        // Register site and control panel events
+        if (Craft::$app->getRequest()->getIsSiteRequest()) {
+            $this->registerHintsUtilityEvents();
+        } else {
             $this->registerCpUrlRules();
             $this->registerUtilities();
             $this->registerWidgets();
@@ -305,7 +309,7 @@ class Blitz extends Plugin
      */
     private function registerCacheableRequestEvents(): void
     {
-        // Register application init event
+        // Register web application init event
         Event::on(Application::class, Application::EVENT_INIT,
             function() {
                 $this->cacheRequest->setDefaultCacheControlHeader();
@@ -405,57 +409,6 @@ class Blitz extends Plugin
     }
 
     /**
-     * Registers resave element events.
-     *
-     * Using Craft’s bulk operations events is not possible, since we need to track changes and the bulk operation can span multiple requests.
-     * https://craftcms.com/docs/5.x/extend/events.html#bulk-operations
-     */
-    private function registerResaveElementEvents(): void
-    {
-        // Enable batch mode
-        $events = [
-            [Elements::class, Elements::EVENT_BEFORE_RESAVE_ELEMENTS],
-            [Elements::class, Elements::EVENT_BEFORE_PROPAGATE_ELEMENTS],
-        ];
-        foreach ($events as $event) {
-            Event::on($event[0], $event[1],
-                function() {
-                    $this->refreshCache->batchMode = true;
-                }
-            );
-        }
-
-        Event::on(Queue::class, Queue::EVENT_BEFORE_EXEC,
-            function(ExecEvent $event) {
-                if ($event->job instanceof ResaveElements) {
-                    $this->refreshCache->batchMode = true;
-                }
-            }
-        );
-
-        // Refresh the cache
-        $events = [
-            [Elements::class, Elements::EVENT_AFTER_RESAVE_ELEMENTS],
-            [Elements::class, Elements::EVENT_AFTER_PROPAGATE_ELEMENTS],
-        ];
-        foreach ($events as $event) {
-            Event::on($event[0], $event[1],
-                function() {
-                    $this->refreshCache->refresh();
-                }
-            );
-        }
-
-        Event::on(CraftQueue::class, CraftQueue::EVENT_AFTER_EXEC_AND_RELEASE,
-            function(ExecEvent $event) {
-                if ($event->job instanceof ResaveElements) {
-                    $this->refreshCache->refresh();
-                }
-            }
-        );
-    }
-
-    /**
      * Registers structure events
      */
     private function registerStructureEvents(): void
@@ -488,16 +441,10 @@ class Blitz extends Plugin
      */
     private function registerHintsUtilityEvents(): void
     {
-        // Ignore CP requests.
-        if (Craft::$app->getRequest()->getIsCpRequest()) {
-            return;
-        }
-
         if (!$this->settings->hintsEnabled) {
             return;
         }
 
-        // Add the event listener without “appending”, so it triggers as early as possible.
         Event::on(ElementQuery::class, ElementQuery::EVENT_BEFORE_PREPARE,
             function(CancelableEvent $event) {
                 /** @var ElementQuery $elementQuery */