+++ src/config.php
@@ -153,7 +153,6 @@ return [
         // `\putyourlightson\blitz\models\SettingsModel::QUERY_STRINGS_CACHE_URLS_AS_SAME_PAGE`: Cache URLs with query strings as the same page
         //'queryStringCaching' => \putyourlightson\blitz\models\SettingsModel::QUERY_STRINGS_DO_NOT_CACHE_URLS,
 
-
         // The query string parameters to include (retain) when caching a URL (regular expressions may be used).
         //'includedQueryStringParams' => [
         //    [
@@ -195,7 +194,7 @@ return [
         // Whether the cache should be refreshed when an element is saved but not live.
         //'refreshCacheWhenElementSavedNotLive' => false,
 
-        // Whether non-HTML responses should be cached. With this setting enabled, Blitz will also cache pages that return non-HTML responses. If enabled, you should ensure that URIs that should not be caches, such as API endpoints, XML sitemaps, etc. are added as excluded URI patterns.
+        // Whether non-HTML responses should be cached. With this setting enabled, Blitz will also cache pages that return non-HTML responses. If enabled, you should ensure that URIs that should not be cached, such as API endpoints, XML sitemaps, etc. are added as excluded URI patterns.
         //'cacheNonHtmlResponses' => false,
 
         // Whether elements should be tracked in the database.
@@ -226,7 +225,6 @@ return [
         // The integrations to initialise.
         //'integrations' => [
         //    'putyourlightson\blitz\drivers\integrations\CommerceIntegration',
-        //    'putyourlightson\blitz\drivers\integrations\FeedMeIntegration',
         //    'putyourlightson\blitz\drivers\integrations\SeomaticIntegration',
         //],
 
@@ -234,10 +232,10 @@ return [
         //'defaultCacheControlHeader' => 'no-store',
 
         // The value to send in the cache control header for cached pages.
-        //'cacheControlHeader' => 'public, maxage=31536000',
+        //'cacheControlHeader' => 'public, s-maxage=31536000, max-age=0',
 
         // The value to send in the cache control header for expired pages.
-        //'cacheControlHeaderExpired' => 'public, maxage=5',
+        //'cacheControlHeaderExpired' => 'public, s-maxage=5, max-age=0',
 
         // Whether an `X-Powered-By: Blitz` header should be added to the response.
         //'sendPoweredByHeader' => true,