+++ src/drivers/integrations/FeedMeIntegration.php
@@ -5,10 +5,9 @@
 
 namespace putyourlightson\blitz\drivers\integrations;
 
-use craft\feedme\services\Process;
-use putyourlightson\blitz\Blitz;
-use yii\base\Event;
-
+/**
+ * @deprecated in 5.7.0
+ */
 class FeedMeIntegration extends BaseIntegration
 {
     /**
@@ -19,23 +18,5 @@ class FeedMeIntegration extends BaseIntegration
         return [
             ['handle' => 'feed-me', 'version' => '4.0.0'],
         ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public static function registerEvents(): void
-    {
-        Event::on(Process::class, Process::EVENT_BEFORE_PROCESS_FEED,
-            function() {
-                Blitz::$plugin->refreshCache->batchMode = true;
-            }
-        );
-
-        Event::on(Process::class, Process::EVENT_AFTER_PROCESS_FEED,
-            function() {
-                Blitz::$plugin->refreshCache->refresh();
-            }
-        );
     }
 }