+++ src/services/RefreshCacheService.php
@@ -36,10 +36,6 @@ use yii\db\ActiveQuery;
 
 /**
  * This class is responsible for keeping the cache fresh.
- * When one or more cacheable element are updated, they are added to the `$elements` array.
- * If `$batchMode` is false then the `refresh()` method is called immediately,
- * otherwise it is triggered by a resave event.
- * The `refresh()` method creates a `RefreshCacheJob` so that refreshing happens asynchronously.
  *
  * @property SiteUriModel[] $allSiteUris
  */
@@ -97,8 +93,15 @@ class RefreshCacheService extends Component
 
     /**
      * @var bool
+     * @deprecated in 5.7.0
      */
     public bool $batchMode = false;
+
+    /**
+     * @var bool Whether the cache should be forcibly generated when refreshed.
+     * @since 5.7.0
+     */
+    public bool $forceGenerate = false;
 
     /**
      * @var RefreshDataModel
@@ -215,11 +218,6 @@ class RefreshCacheService extends Component
 
         if ($this->hasEventHandlers(self::EVENT_AFTER_ADD_ELEMENT)) {
             $this->trigger(self::EVENT_AFTER_ADD_ELEMENT, $event);
-        }
-
-        // If batch mode is on then the refresh will be triggered later
-        if ($this->batchMode === false) {
-            $this->refresh();
         }
     }
 
@@ -346,6 +344,8 @@ class RefreshCacheService extends Component
         if ($this->refreshData->isEmpty()) {
             return;
         }
+
+        $forceGenerate = $forceGenerate || $this->forceGenerate;
 
         $job = new RefreshCacheJob([
             'data' => $this->refreshData->data,
@@ -496,8 +496,6 @@ class RefreshCacheService extends Component
      */
     public function refreshExpiredCache(): void
     {
-        $this->batchMode = true;
-
         $cacheIds = Blitz::$plugin->expireCache->getExpiredCacheIds();
         $this->addCacheIds($cacheIds);
         $this->addExpiredElements();
@@ -512,11 +510,7 @@ class RefreshCacheService extends Component
      */
     public function refreshExpiredElements(): void
     {
-        $this->batchMode = true;
-
         $this->addExpiredElements();
-
-        $this->refresh();
     }
 
     /**
@@ -549,8 +543,6 @@ class RefreshCacheService extends Component
 
         $cacheIds = Blitz::$plugin->cacheTags->getCacheIds($tags);
         $this->addCacheIds($cacheIds);
-
-        $this->refresh();
 
         if ($this->hasEventHandlers(self::EVENT_AFTER_REFRESH_CACHE_TAGS)) {
             $this->trigger(self::EVENT_AFTER_REFRESH_CACHE_TAGS, $event);
@@ -592,12 +584,10 @@ class RefreshCacheService extends Component
     }
 
     /**
-     * Refreshes the cache, forcibly generating the cache if it will not be cleared.
+     * Causes the cache to be forcibly generated if it will not be cleared.
      */
     private function forceRefresh(): void
     {
-        $forceGenerate = !Blitz::$plugin->settings->shouldClearOnRefresh();
-
-        $this->refresh(false, $forceGenerate);
+        $this->forceGenerate = !Blitz::$plugin->settings->shouldClearOnRefresh();
     }
 }