+++ tests/pest/Feature/HintsTest.php
@@ -5,7 +5,6 @@
  */
 
 use craft\elements\db\ElementQuery;
-use craft\elements\Entry;
 use craft\events\CancelableEvent;
 use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\models\HintModel;
@@ -38,18 +37,66 @@ beforeEach(function() {
     );
 });
 
-test('Hint is recorded for a related element query that is lazy-loaded', function() {
-    Entry::find()->section('single')->one()->relatedTo->all();
+test('Hint is recorded for a matrix element query that is lazy-loaded', function() {
+    getSingleEntry()->matrix->all();
     Blitz::$plugin->hints->save();
 
     expect(HintRecord::find()->count())
         ->toBe(1);
 });
 
-test('Hint is not recorded for a related element query that is lazy eager-loaded', function() {
-    Entry::find()->section('single')->one()->relatedTo->eagerly()->all();
+test('Hint is not recorded for a matrix element query that is lazy eager-loaded', function() {
+    getSingleEntry()->matrix->eagerly()->all();
     Blitz::$plugin->hints->save();
 
     expect(HintRecord::find()->count())
-        ->toEqual(0);
+        ->toBe(0);
+});
+
+test('Hint is recorded for a related entry query that is lazy-loaded', function() {
+    getSingleEntry()->relatedTo->all();
+    Blitz::$plugin->hints->save();
+
+    expect(HintRecord::find()->count())
+        ->toBe(1);
+});
+
+test('Hint is not recorded for a related entry query that is lazy eager-loaded', function() {
+    getSingleEntry()->relatedTo->eagerly()->all();
+    Blitz::$plugin->hints->save();
+
+    expect(HintRecord::find()->count())
+        ->toBe(0);
+});
+
+test('Hint is recorded for a related category query that is lazy-loaded', function() {
+    getSingleEntry()->categories->all();
+    Blitz::$plugin->hints->save();
+
+    expect(HintRecord::find()->count())
+        ->toBe(1);
+});
+
+test('Hint is not recorded for a related category query that is lazy eager-loaded', function() {
+    getSingleEntry()->categories->eagerly()->all();
+    Blitz::$plugin->hints->save();
+
+    expect(HintRecord::find()->count())
+        ->toBe(0);
+});
+
+test('Hint is not recorded for element queries with reference tags', function() {
+    getSingleEntry()->relatedTo->ref('{entry:1@1:url}')->all();
+    Blitz::$plugin->hints->save();
+
+    expect(HintRecord::find()->count())
+        ->toBe(0);
+});
+
+test('Hint is not recorded for parsed reference tags', function() {
+    Craft::$app->getElements()->parseRefs('{entry:1@1:url}');
+    Blitz::$plugin->hints->save();
+
+    expect(HintRecord::find()->count())
+        ->toBe(0);
 });