+++ tests/pest/Helpers.php
@@ -39,6 +39,11 @@ function getChannelSectionId(): int
     return Craft::$app->entries->getSectionByHandle(App::env('TEST_CHANNEL_SECTION_HANDLE'))->id;
 }
 
+function getSingleEntry(): Entry|null
+{
+    return Entry::find()->section('single')->one();
+}
+
 function integrationIsActive(string $class): bool
 {
     return in_array($class, IntegrationHelper::getActiveIntegrations());
@@ -57,11 +62,8 @@ function createSiteUri(int $siteId = 1, string $uri = 'page'): SiteUriModel
     ]);
 }
 
-function createEntry(bool $enabled = true, array $params = [], array $customFields = [], bool $batchMode = false): Entry
+function createEntry(bool $enabled = true, array $params = [], array $customFields = []): Entry
 {
-    $originalBatchMode = Blitz::$plugin->refreshCache->batchMode;
-    Blitz::$plugin->refreshCache->batchMode = $batchMode;
-
     /** @var Entry $entryFactory */
     $entryFactory = EntryFactory::factory();
     foreach ($params as $key => $value) {
@@ -82,7 +84,6 @@ function createEntry(bool $enabled = true, array $params = [], array $customFiel
 
     Blitz::$plugin->generateCache->reset();
     Blitz::$plugin->refreshCache->reset();
-    Blitz::$plugin->refreshCache->batchMode = $originalBatchMode;
 
     return $entry;
 }
@@ -125,10 +126,8 @@ function createAsset(): Asset
     return $asset;
 }
 
-function createProductVariantOrder(bool $batchMode = false, bool $inventoryTracked = false): array
+function createProductVariantOrder(bool $inventoryTracked = false): array
 {
-    $originalBatchMode = Blitz::$plugin->refreshCache->batchMode;
-    Blitz::$plugin->refreshCache->batchMode = $batchMode;
     $faker = FakerFactory::create();
 
     $type = Plugin::getInstance()->productTypes->getProductTypeByHandle(App::env('TEST_PRODUCT_TYPE_HANDLE'));
@@ -162,7 +161,6 @@ function createProductVariantOrder(bool $batchMode = false, bool $inventoryTrack
 
     Blitz::$plugin->generateCache->reset();
     Blitz::$plugin->refreshCache->reset();
-    Blitz::$plugin->refreshCache->batchMode = $originalBatchMode;
 
     return [$variant, $order];
 }