+++ tests/pest/Integration/CommerceTest.php
@@ -8,21 +8,13 @@ use craft\commerce\elements\Order;
 use Mockery\MockInterface;
 use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\drivers\integrations\CommerceIntegration;
-use putyourlightson\blitz\services\RefreshCacheService;
 
 beforeEach(function() {
-    Blitz::$plugin->set('refreshCache', Mockery::mock(RefreshCacheService::class . '[refresh,refreshAll]'));
     Blitz::$plugin->refreshCache->reset();
-    Blitz::$plugin->refreshCache->batchMode = false;
 })->skip(fn() => !integrationIsActive(CommerceIntegration::class), 'Commerce integration not found in active integrations.');
 
 test('Variant with inventory is refreshed on order completion', function() {
-    /** @var MockInterface $refreshCache */
-    $refreshCache = Blitz::$plugin->refreshCache;
-    $refreshCache->shouldReceive('refresh')->once();
-    $refreshCache->shouldNotReceive('refreshAll');
-
-    [$variant, $order] = createProductVariantOrder(batchMode: true, inventoryTracked: true);
+    [$variant, $order] = createProductVariantOrder(inventoryTracked: true);
     $order->trigger(Order::EVENT_AFTER_COMPLETE_ORDER);
 
     expect(Blitz::$plugin->refreshCache->refreshData->getElementIds($variant::class))
@@ -31,11 +23,7 @@ test('Variant with inventory is refreshed on order completion', function() {
 
 test('Variant without inventory is not refreshed on order completion', function() {
     /** @var MockInterface $refreshCache */
-    $refreshCache = Blitz::$plugin->refreshCache;
-    $refreshCache->shouldNotReceive('refresh');
-    $refreshCache->shouldNotReceive('refreshAll');
-
-    [$variant, $order] = createProductVariantOrder(batchMode: true);
+    [$variant, $order] = createProductVariantOrder();
     $order->trigger(Order::EVENT_AFTER_COMPLETE_ORDER);
 
     expect(Blitz::$plugin->refreshCache->refreshData->getElementIds($variant::class))